Updated the project to be compatible with Swift 3 and Xcode 8.
Also, added a test to use Core Foundation level APIs instead of POSIX.

To toggle, adjust the useFoundation variable to true or false.

With iPad Pros, the POSIX issue of not activating the antenna is observable. I can register 3-4 second delays, but when using Core Foundation APIs the delay is typically 0.4 seconds, and never beyond 1 sec in my testing.